test(#875): extract the wide-scan biome oracle into a test util
Move getNearestBaseIDByWideScan out of the biome-field suite into its own
test-utils module, per the convention that helpers do not live inside a test
file.
