+++ libs/game/worldmap-core/src/build-biome-field.test.ts
@@ -1,9 +1,8 @@
 import { expect, test } from 'bun:test';
 import { buildBiomeField } from './build-biome-field';
-import { buildCellNode } from './build-cell-node';
 import { buildRevealDistanceField } from './build-reveal-distance-field';
 import { getBiome } from './get-biome';
-import { toHexPosition } from './to-hex-position';
+import { getNearestBaseIDByWideScan } from './test-utils/get-nearest-base-id-by-wide-scan';
 
 test('it lays texels out on the same grid as the reveal distance field for the same viewport and resolution', () => {
   const viewport = { maxCX: 4, maxCY: 4, minCX: -4, minCY: -4 };
@@ -85,25 +84,3 @@ test('it matches a wide brute-force nearest-node scan at the current jitter', ()
     }
   }
 });
-
-function getNearestBaseIDByWideScan(seed: number, cx: number, cy: number): number {
-  const [sceneX, sceneY] = toHexPosition(cx, cy);
-  let nearestBaseID = -1;
-  let nearestDistanceSq = Number.POSITIVE_INFINITY;
-
-  for (let dy = -3; dy <= 3; dy++) {
-    for (let dx = -3; dx <= 3; dx++) {
-      const cellX = Math.round(cx) + dx;
-      const cellY = Math.round(cy) + dy;
-      const [nodeX, nodeY] = buildCellNode(seed, cellX, cellY).position;
-      const distanceSq = (nodeX - sceneX) ** 2 + (nodeY - sceneY) ** 2;
-
-      if (distanceSq < nearestDistanceSq) {
-        nearestDistanceSq = distanceSq;
-        nearestBaseID = getBiome(seed, cellX, cellY).baseID;
-      }
-    }
-  }
-
-  return nearestBaseID;
-}
